Rename [animation] to [pxAnimation]. Break out [pxObject].  Tidy up.
